fix: :truck: Moved block movement process to movement
Issue: #39
